refactor: move error implementations to backend package
