+++ packages/backend/src/modules/users/index.ts
@@ -1,1 +1,2 @@
 export { UsersModule } from './UsersModule';
+export * from './usersExceptions';
