+++ packages/backend/src/shared/errors/errors.ts
@@ -16,3 +16,25 @@ export abstract class CustomError extends Error {
     }
   }
 }
+
+export abstract class BaseError extends CustomError {}
+
+export abstract class BaseException extends CustomError {}
+
+export class ValidationError extends BaseError {
+  constructor() {
+    super('ValidationError');
+  }
+}
+
+export class ServerError extends BaseError {
+  constructor() {
+    super('ServerError');
+  }
+}
+
+export class ClientError extends BaseError {
+  constructor() {
+    super('ClientError');
+  }
+}
