+++ packages/backend/src/shared/errors/index.ts
@@ -1,0 +1,2 @@
+export { errorHandler } from './errorHandler';
+export * from './errors';
