+++ packages/backend/tests/e2e/getPosts.test.ts
@@ -5,9 +5,9 @@ import supertest from 'supertest';
 import { Post } from '@prisma/client';
 import { DatabaseFixtures } from '../support/fixtures/DatabaseFixtures';
 import { UserInputBuilder } from '@dddforum/shared/tests/support/builders/UserInputBuilder';
-import { ClientError } from '@dddforum/shared/src/errors/errors';
 import { Server } from 'http';
 import { CompositionRoot } from '../../src/core';
+import { ClientError } from '../../src/shared';
 
 const feature = loadFeature(path.join(sharedTestRoot, 'features/getPosts.feature'));
